refactor after RT meeting
