+++ robot_log_visualizer/__main__.py
@@ -5,13 +5,10 @@
 # Released under the terms of the BSD 3-Clause License
 
 import sys
-import os
 
 # GUI
 from robot_log_visualizer.ui.gui import RobotViewerMainWindow
 from PyQt5.QtWidgets import QApplication
-
-from robot_log_visualizer.file_reader.signal_provider import SignalProvider
 
 # Meshcat
 from robot_log_visualizer.robot_visualizer.meshcat_provider import MeshcatProvider
@@ -24,11 +21,8 @@ def main():
         "plot_animation": 0.03,
     }
 
-    # instantiate device_manager
-    signal_provider = SignalProvider(period=thread_periods["signal_provider"])
-
     meshcat_provider = MeshcatProvider(
-        period=thread_periods["meshcat_provider"], signal_provider=signal_provider
+        period=thread_periods["meshcat_provider"]
     )
 
     # instantiate a QApplication
@@ -36,16 +30,13 @@ def main():
 
     # instantiate the main window
     gui = RobotViewerMainWindow(
-        signal_provider=signal_provider,
+        signal_provider_period=thread_periods["signal_provider"],
         meshcat_provider=meshcat_provider,
         animation_period=thread_periods["plot_animation"],
     )
 
     # show the main window
     gui.show()
-
-    signal_provider.start()
-    meshcat_provider.start()
 
     return app.exec_()
