+++ robot_log_visualizer/ui/plot_item.py
@@ -9,12 +9,15 @@ from robot_log_visualizer.plotter.matplotlib_viewer_canvas import MatplotlibView
 
 
 class PlotItem(QFrame):
-    def __init__(self, signal_provider, period):
+    def __init__(self, period):
         super().__init__(None)
         self.ui = Ui_PlotTab()
         self.ui.setupUi(self)
 
         self.canvas = MatplotlibViewerCanvas(
-            parent=self, period=period, signal_provider=signal_provider
+            parent=self, period=period
         )
         self.ui.plotLayout.addWidget(self.canvas)
+
+    def set_signal_provider(self, signal_provider):
+        self.canvas.set_signal_provider(signal_provider)
